+++ CHANGELOG.md
@@ -8,6 +8,20 @@ file.
 
 ### Changed
 
+## [0.20.0] 2022-03-20
+### Added
+
+### Changed
+- Stop adding `LD_LIBRARY_PATHS` to process env twice
+- [Internal] can now run `cargo test` on tarpaulin without need of `--test-threads 1`
+- Force --test-threads 1 for --follow-exec unless there's `--implicit-test-threads`
+- Add markers to event log to show where state machine iterations start and end, fix fork parent tracing
+- Handle exec following in vfork children
+- Continue vfork parents so test execution isn't stalled when tracing children
+- Make `--forward` default signal behaviour
+- Fix follow-exec aliasing for config file
+- Fix `force_clean` merging to take into account the default being true
+
 ## [0.19.1] 2022-01-16
 ### Added
 - Added support for `RUST_TEST_THREADS` to specify test threads instead of --test-threads